requirements.txt
streamlit
SpeechRecognition
pydub
python-dotenv
google-generativeai